feat: remove file with if

+++ 04_copiando_e_movendo_arquivos/03_deletando_arquivo.py
@@ -8,5 +8,6 @@ print(pasta_atual)
 caminho_arquivo_origem = pasta_atual / 'assets' / 'file_deleted.txt'
 print(caminho_arquivo_origem)
 caminho_pasta_destino = pasta_atual / 'assets/destino/03/file_deleted.txt'
-shutil.copyfile(caminho_arquivo_origem, caminho_pasta_destino)
-os.remove(caminho_arquivo_origem)
+if caminho_arquivo_origem.exists():
+  shutil.copyfile(caminho_arquivo_origem, caminho_pasta_destino)
+  os.remove(caminho_arquivo_origem)
